Merge with upstream

+++ src/PThreads.cpp
@@ -1,3 +1,9 @@
+/*
+ * Copyright 2002, Log4cpp Project. All rights reserved.
+ *
+ * See the COPYING file for the terms of usage and distribution.
+ */
+
 #include <log4cpp/threading/Threading.hh>
 
 #if defined(LOG4CPP_HAVE_THREADING) && defined(LOG4CPP_USE_PTHREADS)
@@ -6,13 +12,13 @@ namespace log4cpp {
     namespace threading {
 
         char* getThreadId(char* buffer) {
-            ::sprintf(buffer, "%ld", (long int)pthread_self());
+            ::sprintf(buffer, "%lu", (long int)pthread_self());
             return buffer;
         }
 
         std::string getThreadId() {
             char buffer[16];
-            ::sprintf(buffer, "%ld", (long int)pthread_self());
+            ::sprintf(buffer, "%lu", pthread_self());	// thread id unsigned
             return std::string(buffer);     
         }
